+++ projects/rocprofiler-sdk/tests/rocprofv3/advanced-thread-trace/validate.py
@@ -201,28 +201,6 @@ def read_att_stats_kernel_names(output_path):
     return traced_kernel_names
 
 
-def test_counter_collection_with_att(json_data):
-    data = json_data["rocprofiler-sdk-tool"]
-    assert data["strings"]["att_filenames"]
-    counter_names = {
-        counter["id"]["handle"]: counter["name"] for counter in data["counters"]
-    }
-    callbacks = data["callback_records"]["counter_collection"]
-    assert callbacks
-
-    found_positive_value = False
-    for entry in callbacks:
-        dispatch = entry["dispatch_data"]
-        assert dispatch["dispatch_info"]["dispatch_id"] > 0
-        assert dispatch["end_timestamp"] >= dispatch["start_timestamp"]
-        assert entry["records"]
-        for record in entry["records"]:
-            assert counter_names[record["counter_id"]["handle"]] == "SQ_WAVES"
-            assert record["value"] >= 0
-            found_positive_value = found_positive_value or record["value"] > 0
-    assert found_positive_value
-
-
 def test_json_data(json_data):
     data = json_data["rocprofiler-sdk-tool"]
     strings = data["strings"]
